% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_extract.R
\name{gee_extract}
\alias{gee_extract}
\title{Extract values of Images or ImageCollections through Google Earth Engine}
\usage{
gee_extract(x, image_name, reducer_name = "ee.Reducer.mean()",
  reducer_scale = 30, chunksize = 2000, n_cores = NULL)
}
\arguments{
\item{x}{Locations (sf or sp object) used for extraction}

\item{image_name}{Name of Image or ImageCollection}

\item{reducer_name}{Name of reducer}

\item{reducer_scale}{Scale of reducer in meters}

\item{chunksize}{Size of chunks in which data is split to allow extracting values for large input datasets}

\item{n_cores}{Number of cores to use for sending extraction calls in parallel; Default (NULL) indicates  no parallelization}
}
\value{
A sf/Spatial object with extracted values added
}
\description{
Extract values of Images or ImageCollections through Google Earth Engine
}
