% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_match_extract.R
\name{gee_match_extract}
\alias{gee_match_extract}
\title{Extract values based on temporal matching finding closest Image in an ImageCollections through Google Earth Engine}
\usage{
gee_match_extract(x, collection_name, date_col = "date",
  reducer_name = "ee.Reducer.mean()", reducer_scale = 30,
  chunksize = 2000, n_cores = NULL, prefilter = FALSE,
  prefilter_buffer = 20)
}
\arguments{
\item{x}{Locations (sf or sp object) used for extraction}

\item{collection_name}{Name of ImageCollection}

\item{date_col}{Name of date column (Required format: yyyy-mm-dd) used for matching; Default is 'date'}

\item{reducer_name}{Name of reducer}

\item{reducer_scale}{Scale of reducer in meters}

\item{chunksize}{Size of chunks in which data is split to allow extracting values for large input datasets}

\item{n_cores}{Number of cores to use for sending extraction calls in parallel; Default (NULL) indicates  no parallelization}

\item{prefilter}{Whether to pre-filter the ImageCollection. Useful to speed up processing.}

\item{prefilter_buffer}{How many days +- the target date should be included in pre-filtering?}
}
\value{
A sf/Spatial object with extracted values added
}
\description{
Extract values based on temporal matching finding closest Image in an ImageCollections through Google Earth Engine
}
